Fixed 64-bit bug that prevented NumArrays from being bigger than 4Gb. Better error reporting from pipeliner, return proper error codes in many cases.

+++ DMI/src/NumArray.cc
@@ -320,7 +320,7 @@ void DMI::NumArray::init (const LoShape & shape,int flags,TypeId realtype)
   itsWritableRef = false;    
   itsShape = shape;
   itsSize  = shape.product();
-  int sz = sizeof(NumArray::Header) + sizeof(uint)*shape.size();
+  size_t sz = sizeof(NumArray::Header) + sizeof(uint)*shape.size();
   // Align data on 8 bytes.
   itsDataOffset = (sz+7) / 8 * 8;
   sz = itsDataOffset + itsSize*itsElemSize;
